requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-localflavor==3.0.1
python-stdnum==1.14
pytz==2020.1
sqlparse==0.4.0
validate-docbr==1.7.0
djangorestframework==3.12.1
psycopg2-binary>=2.8
django-heroku>=0.3.1
celery>=4.3.0
pylint-celery==0.3
django-celery-beat>=2.2.0
psycopg2>=2.8.6
gunicorn>=20.0.4
